Screen redraw with control+l
If you make the terminal very small and play sometimes the display glitches and creates phantom tiles below the board. I just added an option to redraw the screen when ctrl+l is pressed so the fake squares would disappear. 
I used "\e[1;1H\e[2J" to clear the screen instead of "\033[?25h\033[m" because otherwise the cursor would appear on the bottom left.